requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9
plotly==5.18.0
python-dotenv==1.0.0